# backend/docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Provider": {
        "enum": [
          "discogs",
          "ebay",
          "mock",
          "musicbrainz",
          "spotify"
        ],
        "title": "Provider",
        "type": "string"
      },
      "ProviderRequestAdminOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "endpoint",
          "method",
          "status_code",
          "duration_ms",
          "error",
          "meta",
          "created_at",
          "id",
          "user_id"
        ],
        "title": "ProviderRequestAdminOut",
        "type": "object"
      },
      "ProviderRequestOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          }
        },
        "required": [
          "provider",
          "endpoint",
          "method",
          "status_code",
          "duration_ms",
          "error",
          "meta",
          "created_at"
        ],
        "title": "ProviderRequestOut",
        "type": "object"
      },
      "ProviderRequestSummaryOut": {
        "properties": {
          "avg_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Duration Ms"
          },
          "error_requests": {
            "title": "Error Requests",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "total_requests": {
            "title": "Total Requests",
            "type": "integer"
          }
        },
        "required": [
          "provider",
          "total_requests",
          "error_requests",
          "avg_duration_ms"
        ],
        "title": "ProviderRequestSummaryOut",
        "type": "object"
      },
      "SaveSearchAlertRequest": {
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "poll_interval_seconds": {
            "default": 600,
            "maximum": 86400.0,
            "minimum": 30.0,
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "query": {
            "$ref": "#/components/schemas/SearchQuery"
          }
        },
        "required": [
          "name",
          "query"
        ],
        "title": "SaveSearchAlertRequest",
        "type": "object"
      },
...
      "SearchPagination": {
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "returned": {
            "title": "Returned",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "returned",
          "total_pages",
          "has_next"
        ],
        "title": "SearchPagination",
        "type": "object"
      },
      "SearchQuery": {
        "properties": {
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "max_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Price"
          },
          "min_condition": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Condition"
          },
          "min_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Price"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 24,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "providers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providers"
          }
        },
        "title": "SearchQuery",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchListingOut"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/SearchPagination"
          },
          "provider_errors": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Provider Errors",
            "type": "object"
          },
          "providers_searched": {
            "items": {
              "type": "string"
            },
            "title": "Providers Searched",
            "type": "array"
          }
        },
        "required": [
          "items",
          "pagination",
          "providers_searched"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WatchRuleOut": {
        "example": {
          "created_at": "2026-01-20T11:52:00+00:00",
          "id": "80dc6333-3c3c-49b8-a803-938783fbeb99",
          "is_active": true,
          "last_run_at": "2026-01-20T12:00:00+00:00",
          "name": "Rare techno under $40",
          "next_run_at": "2026-01-20T12:10:00+00:00",
          "poll_interval_seconds": 600,
          "query": {
            "max_price": 40,
            "q": "detroit techno",
            "sources": [
              "discogs"
            ]
          },
          "updated_at": "2026-01-20T12:00:00+00:00",
          "user_id": "8f2a5009-c0a2-4f90-8f1b-c1716c26bf06"
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run At"
          },
          "poll_interval_seconds": {
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "query",
          "is_active",
          "poll_interval_seconds",
          "last_run_at",
          "next_run_at",
          "created_at",
          "updated_at"
        ],
        "title": "WatchRuleOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "watch-wax-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/provider-requests": {
      "get": {
        "operationId": "list_provider_requests_api_provider_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderRequestOut"
                  },
                  "title": "Response List Provider Requests Api Provider Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Provider Requests",
        "tags": [
          "provider-requests"
        ]
      }
    },
    "/api/provider-requests/admin": {
      "get": {
        "operationId": "list_provider_requests_admin_api_provider_requests_admin_get",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Provider"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "status_code_gte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 599,
                  "minimum": 100,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Code Gte"
            }
          },
          {
            "in": "query",
            "name": "status_code_lte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 599,
                  "minimum": 100,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Code Lte"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderRequestAdminOut"
                  },
                  "title": "Response List Provider Requests Admin Api Provider Requests Admin Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Provider Requests Admin",
        "tags": [
          "provider-requests"
        ]
      }
    },
    "/api/provider-requests/admin/summary": {
      "get": {
        "operationId": "provider_request_summary_admin_api_provider_requests_admin_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Provider"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "status_code_gte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 599,
                  "minimum": 100,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Code Gte"
            }
          },
          {
            "in": "query",
            "name": "status_code_lte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 599,
                  "minimum": 100,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Code Lte"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderRequestSummaryOut"
                  },
                  "title": "Response Provider Request Summary Admin Api Provider Requests Admin Summary Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Provider Request Summary Admin",
        "tags": [
          "provider-requests"
        ]
      }
    },
    "/api/provider-requests/summary": {
      "get": {
        "operationId": "provider_request_summary_api_provider_requests_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderRequestSummaryOut"
                  },
                  "title": "Response Provider Request Summary Api Provider Requests Summary Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Provider Request Summary",
        "tags": [
          "provider-requests"
        ]
      }
    },
    "/api/search": {
      "post": {
        "operationId": "search_listings_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Listings",
        "tags": [
          "search"
        ]
      }
    },
    "/api/search/save-alert": {
      "post": {
        "operationId": "save_alert_api_search_save_alert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSearchAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Alert",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}